% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-image.R
\name{analyzeImage}
\alias{analyzeImage}
\title{Process an image}
\usage{
analyzeImage(file, griddf, pixelsmm, boxWidth, nBoxGrid, plateRadius,
  thresh = NULL, deltax = 0, deltay = 0, savedir = ".", hires = 1500,
  lowres = 500, homeratio = 1/2, channels = 3, rimThreshold = 0.9,
  minArea = 1.5, rotation = 0, checkrotation = TRUE,
  checklocation = TRUE, doqc = TRUE, verbose = FALSE,
  overwriteQc = FALSE, squareWidth = NULL, minNonEmpty = 0.001,
  makeRelativePath = TRUE, ...)
}
\arguments{
\item{file}{the path of the image to analyze}

\item{griddf}{data frame that describes the grid. Has three
columns, the \code{RANGE} which lists all columns by an integer,
\code{ROW} that lists all rows by a factor and \code{box_num} that
indicates the number of the box (running integer from 1 to number
of boxes). One row per box.}

\item{pixelsmm}{number of pixels that correspond to 1 mm.}

\item{boxWidth}{the width (and height) of a single box in the grid in mm.}

\item{nBoxGrid}{the number of boxes in the grid.}

\item{plateRadius}{the radius of the plate in millimeters.}

\item{thresh}{the threshold to appply for deciding what is plant
(close zero) or background (close to 1)}

\item{deltax}{a shift of the places location in millimeters to
apply in to the right if negative or left if positive}

\item{deltay}{a shift of the plates location in millimeters to
apply in to the upwards (if negative number) or downwards (if
positive number)}

\item{savedir}{directory to save pictures for quality control purpose}

\item{hires}{the resolution to use when calculating plant
areas. The higher the more precise but also slower and more memory
intensive.}

\item{lowres}{the resolution in pixels along the width the
image at which to detect rotation and plate eccentricity. The
higher, the more exact but also severly increases computation
time. Does not affect area calculations as this is always done at
maximum resolution.}

\item{homeratio}{the ratio of area of the box in which a feature
occupies the largest area (home box) and the occupance of the box
the feature occupies second most area (neighbor box). With 1/3,
the home box must account for 3 times as much area as the neighbor
box or the feature will be marked ambiguous and its home will be
subjected to grid-segmentation.}

\item{channels}{the channels to use when converting to grey scale
(R, G, B)}

\item{rimThreshold}{a step of sorting features into different
boxes is to identify features that are too far from any box center
that may be part of the plate edges or detections of the physical
plate itself. Features that are farther away from any box center
than half the box size times this constant are considered bad and
removed.}

\item{minArea}{minimum area of a plant. A box with a total
feature area less than this number will have area set to missing
value.}

\item{rotation}{the rotation of the plate in degrees}

\item{checkrotation}{check that plate is placed with grid
perpendicular to edges and correct the image if it the rotation
angle is more than a given small angle (defined by
\code{\link{optimAngle}})}

\item{checklocation}{check that the plate is in the center of the
image and adjust the image otherwise. Ignored if either deltax or
deltay are set to a non-zero value.}

\item{doqc}{generate quality control picture or not}

\item{verbose}{print some messages about progress}

\item{overwriteQc}{overwrite any already exisiting QC
images.}

\item{squareWidth}{The desired width of the image after initial
cropping. Prior to any further analysis, the image is made square
by cropping the left and right margins assuming that the max
horizontal excentricity of the plate is smaller than top / bottom
margin. Set this argument to zero for very excentric plates in
which case cropping is skipped.}

\item{minNonEmpty}{the minimum fraction of pixels above the threshold for a
plate to be considered empty}

\item{makeRelativePath}{should the path to the qc image be made
relative to the current working directory (setting to false has
expected effect but true is only meaningful when using the built-in
template report system in rosettR).}

\item{...}{passed on to \code{\link{findPlate}} and
\code{\link{optimAngle}}.}
}
\value{
a data frame with information about plant areas, locations
}
\description{
Does all steps from checking plate location, rotating the picture
to ensure grid is horizontal/vertical, gridding, calculating plant
areas and generating quality control image.
}
\examples{
file <- system.file("examples", "plate.jpg", package="rosettR")
meta <- metaTemplate(letters[1:4], LETTERS[1:2])
df <- analyzeImage(file, meta$griddf, 3.7454, 20, 6, 75, verbose=TRUE,
                   checkrotation=FALSE)
library(EBImage)
display(readImage(df$qc_picture[1]), method="raster")
}
\author{
Henning Redestig
}

