% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-image.R
\name{calibrateScale}
\alias{calibrateScale}
\title{Calibrate the scale of an image}
\usage{
calibrateScale(path)
}
\arguments{
\item{path}{path to the experiment to calibrate the scale for}
}
\value{
the number of pixels per mm
}
\description{
Used to figure out how many pixels there are on 1 mm by allowing
the user to click on an image. Updates the meta data with the
calculated pixels per millimeter
}

