% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-image.R
\name{colorWellFrames}
\alias{colorWellFrames}
\title{Color the frame of all wells}
\usage{
colorWellFrames(mat, griddf, wellWidth)
}
\arguments{
\item{mat}{a matrix representing the image}

\item{griddf}{a data frame that defines the grid}

\item{wellWidth}{the width of well in pixels}
}
\value{
the image with frames indicated
}
\description{
Set the perimiters of all wells to 1 or 2 depending on whether they are
indicated as being ambiguous
}

