% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-utils.R
\name{createPlateQcDf}
\alias{createPlateQcDf}
\title{Quality control statistics for plate images}
\usage{
createPlateQcDf(df, html = TRUE)
}
\arguments{
\item{df}{a data frame with all data obtained from a plate
experiment (a \code{phenodata} dataset).}

\item{html}{should the dataframe be formatted to HTML or not}
}
\value{
A data frame with quality control statistics ready for
HTML export (if so requested). Links are provided relatively
assuming that the output file is two levels below the output
directory, e.g., in Output/reports/myanalysis/plate-qc-df.html.
}
\description{
Creates a data frame with QC statistics from the result of the
image processing pipeline
}
\details{
The following components are given \describe{

\item{rotation}{the clockwise rotation in degrees that was applied
to the plate. Missing value indicates that no rotation was
done. An absolute value close to the maximum tested rotation may
indicate poor quality}

\item{eccentricity}{an estimate for how far away the plate
deviates from the center in millimeter. A large value may indicate
poor quality but the plate detection algorithm is unstable an
often over-estimates this value.}

\item{ambiguous\_boxes}{the number of boxes on the plate that had
features that could not be resolved to a specific boxes and hence
treated by hard-splitting of the box and summing features. A
number greater than 0 indicates poor quality.}

\item{max_features}{the maximum number of features in a box. The
larger the number the worse quality.}

}
}
\author{
Henning Redestig
}

