% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-utils.R
\name{dateTaken}
\alias{dateTaken}
\title{The date an image was taken}
\usage{
dateTaken(file)
}
\arguments{
\item{file}{an image file (jpg)}
}
\value{
the date
}
\description{
Read create date from the exif tags using.
}
\examples{
dateTaken(system.file("examples/plate001.jpg", package="rosettR"))
}
\author{
Henning Redestig
}
\references{
Uses Mayank Lahiri's exif parsing library
\url{https://github.com/mayanklahiri/easyexif}
}

