% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-image.R
\name{doProcessPlateImages}
\alias{doProcessPlateImages}
\title{The function that actually process the plate images}
\usage{
doProcessPlateImages(path, mf = readManifest(path), meta = readMeta(path),
  verbose = FALSE, .progress = "text", .parallel = FALSE, ...)
}
\arguments{
\item{path}{the path to the plate experiment to process}

\item{mf}{the manifest - a data frame describing all images to be processed}

\item{meta}{a list with parameters to use for \code{\link{analyzeImage}}, see
\code{\link{metaTemplate}}}

\item{verbose}{if true, display progress bar}

\item{.progress}{passed to \code{\link{ddply}}}

\item{.parallel}{passed to \code{\link{ddply}}}

\item{...}{passed on to \code{analyzeImage}. Named arguments have precedence
over arguments in the \code{meta} list.}
}
\value{
the resulting data frame
}
\description{
Not intended for user-level usage.
}
\author{
Henning Redestig
}

