% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-package.R
\docType{data}
\name{exampleMetadata}
\alias{exampleMetadata}
\title{Example metadata for a plate experiment}
\usage{
data(exampleMetadata)
}
\description{
A simple list representation of meta data for a plate experiment.
}
\author{
Henning Redestig
}
\keyword{datasets}

