% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-utils.R
\name{expandManifest}
\alias{expandManifest}
\title{Expand experiment design}
\usage{
expandManifest(meta, plateName = "plate\%03d.jpg", plateOffset = 0,
  plateOrder, ...)
}
\arguments{
\item{meta}{the meta data object that defines the experiment design; the
number of repeats, timepoints, germplasms and treatments}

\item{plateName}{canonical file name of the plates to use with
\code{\link{sprintf}}.}

\item{plateOffset}{the number that the plate numbering should be
offset to. The number of the first plate is 1 + plateoffset.}

\item{plateOrder}{An pre-defined order of the plate after they
have been randomized. The plate order \emph{must} sort to correct
order when using \code{\link{rank}} but can be either numerical or
character.}

\item{...}{not used}
}
\value{
a data frame representing the suggested design of the
experiment
}
\description{
Given the parameters defined in the metadata, expand the balanced
design matrix with each germplasm and each treatment represented
in each repeat of the experiment.
}
\examples{
data(exampleMetadata)
exampleMetadata
expandManifest(exampleMetadata)
}
\author{
Henning Redestig
}

