% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-image.R
\name{findPlate}
\alias{findPlate}
\title{Find a plate in an image}
\usage{
findPlate(im, radius, threshold)
}
\arguments{
\item{im}{an image matrix}

\item{radius}{the radius of the circle}

\item{threshold}{the threshold pixel value}
}
\value{
a list with optimal deltax and deltay parameters
(deviations from origo)
}
\description{
Uses Nelder-Mead optimization to find the position of a circle
that captures as many pixels below a given threshold as possible.
}
\author{
Henning Redestig
}

