% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-image.R
\name{makeGrey}
\alias{makeGrey}
\title{Make grey scale}
\usage{
makeGrey(im, channels = 3)
}
\arguments{
\item{im}{an image}

\item{channels}{the channels to sum}
}
\value{
the grey scale iamge
}
\description{
Suitable for finding leafs e.g. only use the blue channel in which green
tissue is maximally absorbing
}
\author{
Henning Redestig
}

