% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-ui.R
\name{makeTestExperiment}
\alias{makeTestExperiment}
\title{Create a dummy plate experiment}
\usage{
makeTestExperiment(exdir = ".")
}
\arguments{
\item{exdir}{where to create the test experiment}
}
\value{
invisibly, the path to the created test experiment
}
\description{
To be used for testing purposed in documentation and unit-tests.
}
\examples{
makeTestExperiment(tempdir())
}
\author{
Henning Redestig
}

