% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-image.R
\name{markCircle}
\alias{markCircle}
\title{Mark a circle in a matrix}
\usage{
markCircle(mat, x, y, radius, width = 2, edgeValue = 1, fillValue = NULL)
}
\arguments{
\item{mat}{a numeric matrix}

\item{x}{the x-coordinate of the centre of the circle}

\item{y}{the y-coordinate of the centre of the circle}

\item{radius}{the radius of the circle}

\item{width}{the width of the edge}

\item{edgeValue}{the value to set the edge to}

\item{fillValue}{the value to set the content of the circle to}
}
\value{
the modified data matrix
}
\description{
Set edge and fill of a circle in a matrix to defined values.
}
\author{
Henning Redestig
mat <- matrix(runif(100 * 200, 0, 1), 200, 100)
marked <- markCircle(mat, 20, 50, 10)
EBImage::display(Image(marked), method="raster")
}

