% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-ui.R
\name{metaTemplate}
\alias{metaTemplate}
\title{Create a meta data template}
\usage{
metaTemplate(genotypes, treatments = "control", timepoints = c(11, 14, 16,
  18), nblocks = 10, description = "", pixelsmm = 16,
  name = c("6x6.abcd", "6x6.ab"), plateRadius = 76, nBoxGrid = 6,
  boxWidth = 20, reference)
}
\arguments{
\item{genotypes}{the names of the genotypes used in the experiment.}

\item{treatments}{a character vector listing the applied treatments}

\item{timepoints}{the timepoints (days) when pictures are taken.}

\item{nblocks}{the number of replicates}

\item{description}{a short description of the experiment to
annotate template reports}

\item{pixelsmm}{the number of pixels per mm on a picture (must be
the same for all pictures)}

\item{name}{the name of the experiment type to use. See details.}

\item{plateRadius}{the radius of the plate (max radius of the lid)}

\item{nBoxGrid}{the number of wells in each row and column of a
plate. Layout is assumed to be square but with the corner wells
left-out.}

\item{boxWidth}{the width of a well in millimeter}

\item{reference}{the genotype that is meant to be used as the
reference. Necessary for the 'compare areas' report but can be left
\code{NULL} if not applicable.}
}
\value{
a list with meta data for the experiment
}
\description{
Each experiment is annotated with a set of parameters stored in a
'meta' data object. Parameters to use may differ between experiment
setup and design. In all cases, the basic experiment setup is to
have plates with seedlings that are taken from the growth chamber
for imaging on a pre-defined set of days-after-sowing.
}
\details{
This package is delivered with three different experiment
designs. These are:

\itemize{ \item{6x6.abcd}{Uses a 6x6 plate grid with genotypes
(lines) allocated to regions a, b, c and d that are placed
clockwise around the plate. Total number of genotypes must thus be
divisible by four.}

\item{6x6.ab}{Uses a 6x6 plate grid with genotypes (lines)
allocated to regions a and b that positioned at the upper and lower
half of the plate.}
}
}
\examples{
metaTemplate(c("foo", "bar", "baz", "qux"),
             c("control", "osmotic_stress"), reference="foo")
}

