% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-ui.R
\name{newExperiment}
\alias{newExperiment}
\title{Start a new plate phenotyping experiment}
\usage{
newExperiment(path, meta)
}
\arguments{
\item{path}{the desired path for the experiment}

\item{meta}{a meta data object as created by \code{\link{metaTemplate}}}
}
\value{
nothing, used for the side effect of create a new experiment directory
}
\description{
The experiment designs implemented in this package are managed in
file system directories with pre-defined structure. Use this
function to create a new experiment with sowing instructions and
directories in which to place pictures.
}
\examples{
meta <- metaTemplate(letters[1:4], c("control", "osmotic_stress"), reference="a")
newExperiment(file.path(tempdir(), "test"), meta)
list.files(tempdir())
}

