% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-image.R
\name{optimAngle}
\alias{optimAngle}
\title{Estimate the optimum rotatory correction to an image with objects
in straight rows and columns}
\usage{
optimAngle(im, pxlsmm, boxWidth, nBoxGrid, angles = c(-3, 3), ...)
}
\arguments{
\item{im}{an image object}

\item{pxlsmm}{the number of pixels per millimeter}

\item{boxWidth}{the number of rows and columns in the expected symmetric grid}

\item{nBoxGrid}{the width of a box}

\item{angles}{min and max values for rotation correction}

\item{...}{not used}
}
\value{
the suggested rotation correction in degrees
}
\description{
Applies equidistant location-fixed mixed model to the summed
horizontal and vertical margin pixel sums to a series of rotated
images with the goal to minimize a the common \eqn{\sigma}.
}
\author{
Henning Redestig
}

