% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-package.R
\name{plate-reports}
\alias{plate-reports}
\title{Template report for plate phenotyping}
\description{
This package comes with a set of template reports that can be compiled for an
experiment to rapidly perform quality control and statistical analysis
presented as html based reports.
}

