% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-reporting.R
\name{plateGallery}
\alias{plateGallery}
\title{Make a thumbnail gallery of the input images}
\usage{
plateGallery(path, what = c("raw", "qc"), parallel = FALSE)
}
\arguments{
\item{path}{a resolvable path to the experiment.}

\item{what}{either 'qc' or 'raw' for galleries of qc images or raw
images respectively.}

\item{parallel}{use registered parallel backend for faster
generation of image thumbnails, see \code{\link{llply}}}
}
\value{
nothing, used for its side effect.
}
\description{
Creates a html thumbnailDir gallery of the input images where the
images are arranged after how the will be interpreted by the
renaming facilities and which day they were taken. The images are
clickable and links to the original input images.
}
\examples{
\dontrun{
plateGallery(pathToExperiment, "raw")
}
}
\author{
Henning Redestig
}
\seealso{
reports
}

