% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-image.R
\name{postProcessPlatePhenodata}
\alias{postProcessPlatePhenodata}
\title{Post-process phenodata}
\usage{
postProcessPlatePhenodata(phenodata, minArea = 1.5, ...)
}
\arguments{
\item{phenodata}{a phenodata object to post process}

\item{minArea}{minimum area of a plant. A box with a total
feature area less than this number will have area set to missing
value.}

\item{...}{not used}
}
\value{
phenodata object
}
\description{
Apply post-process steps to the phenodata obtained from
\code{analyzeImage}. Currently this consists of marking all
measurements for plants that at at least one day are not larger
than the specified minimum area as 'removed'.
}
\author{
Henning Redestig
}

