% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-utils.R
\name{processPlateExperiment}
\alias{processPlateExperiment}
\title{Process a plate experiment}
\usage{
processPlateExperiment(path, meta = readMeta(path), rename = TRUE,
  analyze = TRUE, ...)
}
\arguments{
\item{path}{the directory where to find data for this experiment.}

\item{meta}{the metadata object for this experiment. Read from the
experiment's manifest file if missing.}

\item{rename}{if true, rename image files otherwise assume that
images are renamed externally.}

\item{analyze}{analyze the images or not}

\item{...}{passed on to \code{process_plate_images}. Note in
particular the argument \code{wincores} that enables the use of
more than one CPU.}
}
\value{
The resulting phenodata object if \code{analyze} was true,
otherwise \code{NULL}, both invisibly.
## unpack an example experiment to working directory.
}
\description{
After pictures have been placed in their appropriate places in the
experiment folder, this function renames the files to
plate<number>.jpg and updates the manifest file to indicate the
old file names. After this, the experiment is 'sealed' to avoid
renaming the files again. Images are sorted after the day they
were taken as taken from the EXIF tag in the image file.
}
\details{
After processing, images can be analyzed and data analysis reports
compiled.

This function is intended for interactive use. You probably do not
want to use it in programs directly, instead you may want to look
at \code{\link{analyzeImage}} and \code{\link{processPlateImages}}.
}
\examples{
path <- makeTestExperiment(tempdir())
processPlateExperiment(path, analyze=FALSE)
readManifest(path)
\dontrun{
## optionally analyze the pictures as well
processPlateExperiment(path)
}
}
\author{
Henning Redestig
}

