% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-image.R
\name{processPlateImages}
\alias{processPlateImages}
\title{Process a set of plate images}
\usage{
processPlateImages(path, mf = readManifest(path), meta = readMeta(path),
  verbose = FALSE, save = TRUE, ...)
}
\arguments{
\item{path}{the path to the experiment}

\item{mf}{the manifest data frame for the experiment that must
have a column named 'image' that specifies the path to each image
as well as 'timepoint' that specifies the timepoint at which the
image was taken.}

\item{meta}{the meta data associated with the experiment}

\item{verbose}{display progressbar or not. Disabled for parallel
computation as it does not work reliably.}

\item{save}{should the created phenodata object and a csv file
with the final dataset be placed in the output directory of the
experiment or not}

\item{...}{passed on to \code{\link{analyzeImage}} and
\code{\link{doProcessPlateImages}}}
}
\value{
invisibly, the \code{phenodata} object with the results.
}
\description{
Apply \code{analyzeImage} to all images in an experiment.
}
\details{
Parallelization is done using the the plyr package plugin to the
\code{foreach} package. \code{doParallel} is used on Linux but the
multi-core functionality is (deliberately) defunct on
Windows.. Missing required arguments for \code{analyzeImage} are
taken from the loaded configuration.
}
\author{
Henning Redestig
}

