% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-reporting.R
\name{relativeGrowthRate}
\alias{relativeGrowthRate}
\title{Relative growth rate}
\usage{
relativeGrowthRate(x, timepoints)
}
\arguments{
\item{x}{a numeric time series of measurements}

\item{timepoints}{the time points of the measurements}
}
\value{
a numeric vector of relative growth rates
}
\description{
Calculate the RGR of a time-series.
\deqn{RGR = \frac{ln(w_2) - ln(w_1)}{t_2 - t_1}}
}
\details{
Negative or non-finite growth rates are set to NA.
}
\examples{
relativeGrowthRate(1:10, 1:10)
relativeGrowthRate(rep(1, 10), 1:10)
}
\author{
Henning Redestig
}
\references{
Hoffmann, W. A.; Poorter, H. (2002). "Avoiding Bias in
Calculations of Relative Growth Rate". Annals of Botany 90 (1): 37
}

