% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-utils.R
\name{removeBoxes}
\alias{removeBoxes}
\title{Remove boxes}
\usage{
removeBoxes(path, pattern, platere = "plate\\\\s*\\\\d+", remove = TRUE)
}
\arguments{
\item{path}{path to the experiment (i.e. the top-level folder in
which the Output directory is found)}

\item{pattern}{a character vector listing pattern on the form
"[day]/[plateXXX]/[row:column]". All units can be regular
expressions, e.g. use .*/plate11./3:.* to remove all boxes in the
third row of plate 110 to 119 from all days.}

\item{platere}{a regular expression that matches the plate
name. Can be left missing to use a default of plate[optional
spaces][numbers].jpg.}

\item{remove}{remove if true, un-remove if false.}
}
\value{
nothing, writes in the \code{data.csv} file and saves the
\code{phenodata.rda} file for the corresponding experiment
}
\description{
Flag observations in an experiment dataset as removed to inform
downstream analysis that the corresponding data should be ignored.
}
\examples{
makeTestExperiment(tempdir())
path <- file.path(tempdir(), "rosettrTest")
## remove the plant in 3:3 on plate002 from day 18
removeBoxes(path, ".*/plate002/3:3")
## list the removed plants
subset(readPhenodata(path), removed)
}
\author{
Henning Redestig
}

