% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-utils.R
\name{renamePlateImages}
\alias{renamePlateImages}
\title{Rename plate images}
\usage{
renamePlateImages(path, newnames, dry = TRUE, verbose = TRUE,
  subdirs = NULL)
}
\arguments{
\item{path}{the path to the experiment or path to a directory with
images}

\item{newnames}{the desired new names of the files after they have
been ordered by the time they were taken. The newnames should have
the correct jpg suffix. If they do not, a warning is given and the
jpg suffix is added.}

\item{dry}{if true, do simulation only, no files are renamed. If
false, files will be renamed destructively.}

\item{verbose}{print messages about the file renaming action to be
performed}

\item{subdirs}{character vector indicating subset of directories to process
(see example)}
}
\value{
a data frame that indictes the old and the new name. It is
good to save this data frame to be able to revert file renaming
}
\description{
Images uploaded from a digital camera may have file names that
make it difficult to track which image correspond to which
plate. This function can be used to order the files after when
they were taken (as defined in the EXIF tag) and to rename them to
desired file names. This function is destructive, use with care
and always do a dry run first to make sure that the new file names
are the desired ones.
}
\details{
\emph{For typical plate image analysis, you should not use this
function}, but rather let \code{\link{processPlateExperiment}}
take care of all file renaming. This function is only exported in
order to facilitate transition from legacy software.
}
\examples{
path <- makeTestExperiment(file.path(tempdir(), "xyz"))
newnames <- c("plate001.jpg", "plate002.jpg", "plate003.jpg", "plate004.jpg",
              "plate005.jpg", "plate006.jpg")
## simulate renaming all files
renamePlateImages(path, newnames)
## simulate for a single directory
renamePlateImages(path, newnames, subdirs="D11")
## to actually make changes
renaming <- renamePlateImages(path, newnames, dry=FALSE)
list.files(file.path(path, "D11"))
## undo the renaming
with(renaming, mapply(file.rename, file.path(path, subdir, newname),
file.path(path, subdir, intermediate)))
with(renaming, mapply(file.rename, file.path(path, subdir, intermediate),
file.path(path, subdir, image)))
list.files(file.path(path, "D11"))
}
\author{
Henning Redestig
}

