% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-image.R
\name{reprocessPlateImages}
\alias{reprocessPlateImages}
\title{Re-process a (sub)set of previously processed plate images}
\usage{
reprocessPlateImages(path, mf, verbose = FALSE, save = TRUE, ...)
}
\arguments{
\item{path}{the path to the experiment to reprocess}

\item{mf}{a manifest data frame for the selected images.}

\item{verbose}{display progressbar or not. Disabled for parallel
computation as it does not work reliably.}

\item{save}{should the created phenodata object and csv-file with
the complemented dataset be placed in the output directory of the
experiment or not. Previous results for selected images are lost.}

\item{...}{passed on to \code{\link{analyzeImage}} (precedence over
the parameters defined in \code{meta}) and
\code{\link{doProcessPlateImages}}}
}
\value{
invisibly, the complete (all images, not only those that
were re-processed) \code{phenodata} object with the results.
}
\description{
Apply \code{\link{analyzeImage}} to a selected set of images in an
experiment and combine the results with previously computed
results.
}
\examples{
\dontrun{
path <- makeTestExperiment(tempdir())
pda <- processPlateExperiment(path, checklocation=FALSE, checkrotation=FALSE, verbose=TRUE)
## assume we were not happy with settings for first plate, day 18
mf <- readManifest(path)
mf <- subset(mf, timepoint == 18 & plate  == "plate001.jpg")
## reprocess the plate adding rotation check and correction
pda <- reprocessPlateImages(path, mf, checkrotation=TRUE)
}
}
\author{
Henning Redestig
}
\seealso{
\code{\link{processPlateImages}} that has the same arguments and
functionality but is used for all images in the experiment.
}

