% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-package.R
\docType{package}
\name{rosettr}
\alias{rosettr}
\alias{rosettr-package}
\title{Easy phenotyping of plants grown in plates}
\description{
Phenotyping plants grown in plates
}
\details{
\tabular{ll}{
Package: \tab rosettr \cr
Type: \tab Package \cr
Developed since: \tab 2014 \cr
License: \tab GPL \cr
LazyLoad: \tab yes \cr
}
}
\author{
Henning Redestig <henning.red at gmail.com>
}

