% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-utils.R
\name{rwMeta}
\alias{readManifest}
\alias{readMeta}
\alias{readPhenodata}
\alias{rwMeta}
\alias{writeManifest}
\alias{writeMeta}
\alias{writePhenodata}
\title{Read and write meta-data and analysis results}
\usage{
writeMeta(meta, path)

readMeta(path)

readPhenodata(path)

writePhenodata(path, df)

writeManifest(manifest, path)

readManifest(path)
}
\arguments{
\item{meta}{the meta data object}

\item{path}{path to the experiment directory}

\item{df}{a data frame containing analysis results for the given}

\item{manifest}{a data frame containing the manifest (description
of each plate and well for the whole experiment) experiment}
}
\value{
nothing, used for side effect
}
\description{
rosettR saves meta data and analysis results for each experiment in
a special file in the corresponding experiment directory. Use these
function to read/write data to those files in the format expected by
rosettR.
}

