% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-utils.R
\name{simpleAnovaTableGT}
\alias{simpleAnovaTableGT}
\title{Perform ANOVA for the comparisons in a plate experiment}
\usage{
simpleAnovaTableGT(data, reference, responseVariables)
}
\arguments{
\item{data}{a data frame with plate phenotyping results, such as output from
\code{\link{createPlateTestDf}}}

\item{reference}{the reference genotype}

\item{responseVariables}{the response variable(s) to compute the contrasts
statistics for. The pvalues are corrected for multiple comparisons.}
}
\value{
a data frame with one contrast per row
}
\description{
Perform ANOVA for the comparisons in a plate experiment
}
\examples{
# TODO
}
\seealso{
\code{\link{glht}} which is used to make the multiple comparisons
}

