% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-reporting.R
\name{simpleOutlierTest}
\alias{simpleOutlierTest}
\title{Check for strong outliers}
\usage{
simpleOutlierTest(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
logical indicating for each value if it outlier or not
}
\description{
Use the outlier test implemented in \code{boxplot} to mark outliers in a numeric vector.
}
\examples{
simpleOutlierTest(rnorm(100))
}

