% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-utils.R
\name{updateMeta}
\alias{updateMeta}
\title{Update the meta object}
\usage{
updateMeta(..., meta)
}
\arguments{
\item{...}{named arguments}

\item{meta}{the meta object}
}
\value{
the updated meta object
}
\description{
Meta object is a list of parameters to use as defaults for a given
experiment. One can tweak these defaults for e.g. specific image by passing
parametesr as named arguments. This function then updates the meta object
with these tweaks
}
\author{
Henning Redestig
}

