% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-image.R
\name{wellIndex}
\alias{wellIndex}
\title{The indices in a well}
\usage{
wellIndex(griddf, i, wellWidth)
}
\arguments{
\item{griddf}{the data frame that defines the plate grid}

\item{i}{the index of the well to return (row in the grid data frame )}

\item{wellWidth}{the width (in pixels) of the well}
}
\value{
a list of row and column indices
}
\description{
Get the matrix indices for all pixels in a given well
}
\author{
Henning Redestig
}

