% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettr-image.R
\name{calibrateScale}
\alias{calibrateScale}
\title{Calibrate the scale of an image}
\usage{
calibrateScale(path)
}
\arguments{
\item{path}{path to the experiment to calibrate the scale for}
}
\value{
the number of pixels per mm
}
\description{
In order to make results comparable across different experiments,
it is necessary to know how many pixels correspond to 1
millimeter. To record this pixels-to-millimeter conversion, measure
the distance on a plate between two points that you can easily
recognize in the image. On the plates used to test this package, 80
mm correspondss to 4 wells. Then left click with the mouse on the
same two points separated by that distance. The pixels/mm is then
automatically recorded in the meta-data associated with the given
experiment and used in following image analysis.
}
\examples{
\dontrun{
makeTestExperiment("rosettrTest")
calibrateScale("rosettrTest")
}
}

